app/requirements.txt
GitPython>=2.1.11
pandas>=0.23.4
feather-format>=0.4.0
scikit-learn>=0.20.0
celery>=4.2.1
redis>=3.0.1
flask-restplus>=0.12.1
werkzeug>=0.14.1